requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
black==23.3.0
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
flatbuffers==2.0.7
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
html5lib==1.1
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
keras==2.11.0
kiwisolver==1.4.4
langdetect==1.0.9
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mplcursors==0.5.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.24.2
oauthlib==3.2.2
onnx==1.12.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycparser==2.21
pydyf==0.6.0
PyJWT==2.6.0
pyparsing==3.0.9
pypdfium2==3.21.1
pyphen==0.14.0
python-dateutil==2.8.2
python-doctr==0.6.0
pytz==2023.3
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
Shapely==1.8.5.post1
six==1.16.0
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tf2onnx==1.13.0
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.65.0
typeguard==3.0.2
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
zopfli==0.2.2
